feat: Moving 전략
