--- src/main/kotlin/racing/car/Car.kt
@@ -1,3 +1,0 @@
-package racing.car
-
-class Car(override val spec: Int, val location: Int) : Engine
