--- src/main/kotlin/racing/car/EngineStrategy.kt
@@ -1,5 +1,0 @@
-package racing.car
-
-sealed class EngineStrategy {
-    abstract fun run()
-}
