+++ src/main/kotlin/racing/domain/Car.kt
@@ -1,0 +1,3 @@
+package racing.domain
+
+class Car(val movingStrategy: MovingStrategy, val location: Int)
