+++ src/main/kotlin/racing/domain/Engine.kt
@@ -1,4 +1,4 @@
-package racing.car
+package racing.domain
 
 interface Engine {
     val spec: Int
