+++ src/main/kotlin/racing/domain/MovingStrategy.kt
@@ -1,0 +1,5 @@
+package racing.domain
+
+interface MovingStrategy {
+    fun go(): Boolean
+}
