+++ src/main/kotlin/racing/domain/RandomStrategy.kt
@@ -1,0 +1,14 @@
+package racing.domain
+
+import java.util.Random
+
+class RandomStrategy : MovingStrategy {
+    companion object {
+        const val MOVING_POINT = 4
+    }
+
+    override fun go(): Boolean {
+        val random = Random().nextInt(10)
+        return random >= MOVING_POINT
+    }
+}
